% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_page}
\alias{add_page}
\title{Title}
\usage{
add_page(
  page_name,
  page_menu_name = page_name,
  visibility = "hidden",
  backup_dirs = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{open}{}
}
\description{
Title
}
