% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_pages}
\alias{add_pages}
\title{Title}
\usage{
add_pages(
  pages,
  page_menu_names = pages,
  backup_dirs = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{open}{}
}
\description{
Title
}
