% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_ui_element}
\alias{add_ui_element}
\title{Title}
\usage{
add_ui_element(
  page,
  label,
  output_name,
  ui_func,
  section_num,
  use_comment_group,
  use_aspect_radio
)
}
\arguments{
\item{use_aspect_radio}{}
}
\description{
Title
}
