% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_dash.R
\name{mi_dash}
\alias{mi_dash}
\title{Project hook to create a base dashboard app for the NHSBSA MI team}
\usage{
mi_dash(path, package_name, ...)
}
\arguments{
\item{path}{Path to the package}

\item{package_name}{Name of the package (aka the app or dashboard)}

\item{...}{Never used}
}
\description{
This will create an app skeleton as a starting point for new dashboards built
by the Management Information team. Since it is the function itself which is
provided to \code{golem}, should never need to call explicitly - the arguments are
set as required by \code{golem::create_golem}. Even if the arguments are not actually
used in the project hook, they must be available for \code{golem}.
}
